Split the files into multiple components to modularize them, reduce function length, and make the app more scalable.
