+++ src/App.jsx
@@ -3,7 +3,7 @@ import SignUpPage from "./pages/auth/signup";
 import Dashboard from "./pages/dashboard";
 import Landing from "./pages/landing";
 import { Route, BrowserRouter as Router, Routes } from 'react-router-dom';
-import ScraperDetails from "./pages/scraper/scraper";
+import ScraperDetailsOut from "./pages/scraper/scraper";
 import Output from "./pages/scraper/output";
 import AddScraper from "./pages/addScraper";
 
@@ -16,7 +16,7 @@ function App() {
         <Route path="/signup" element={<SignUpPage />} />
         <Route path="/dashboard" element={<Dashboard />} />
         <Route path="/dashboard/add" element={<AddScraper />} />
-        <Route path="/scraper/:id" element={<ScraperDetails />} />
+        <Route path="/scraper/:id" element={<ScraperDetailsOut />} />
         <Route path="/scraper/output/:id" element={<Output />} />
       </Routes>
     </Router>
